ordered_setup dla przetwarzania obrazow
