--- przetwarzanieobrazow/Projekt/Lab02/tempCodeRunnerFile.py
@@ -1,2 +1,0 @@
-_BGR2GRAY) #metoda 1 
-gray_image_2 = cv.imread(image,0) #meto
