+++ usuniete_konta.csv
@@ -5,3 +5,7 @@ id,nazwa
 4,niehyps
 5,serwer
 6,straszak
+7,test1
+8,test2
+9,test3
+10,test4
